Move Coupon Summary creation to a separate public extension to reuse between IPP and POS
